// src/app/models/employee.ts
export class Employee {
        staffId:number = 0; 
        unitId:number =  0;
        firstName: string ='';
        fatherName: string='';
        grandName: string='';
        sex:number = 0;
        religion:string ='';
        birthDate: Date;
        birthPlace: string ='';
        meritialStatusId:number= 0;
        nationalityId:number= 0;
        regionId:number= 0;
        zoneId:number= 0;
        woredaId:number= 0;
        kebeleId:number= 0;
        email: string = '';
        telMobile: string ='';
        telHome: string='';
        telOffDirect1: string ='';
        telOffDirect2: string ='';
        telOffExt1: string ='';
        roomNo:number = 0;
        educationLevel:string='';
        educationField: string = '';
        positionId:number = 0;
        salary:number = 0;
        empDate: Date ;
        empLetter:string='';
        empStatus:number = 0;
        currentStatus:number= 0;
        terminationDate: Date;
        terminationReason: string = '';
        terminationLetter: string = '';
        photo: string
}